% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{som_basemap}
\alias{som_basemap}
\title{create basemap with settlement points}
\usage{
som_basemap()
}
\value{
a leaflet object with points created from the somalia sampling frame
}
\description{
create basemap with settlement points
}
