% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{som_map_add_points}
\alias{som_map_add_points}
\title{add points of an assesment to the web map}
\usage{
som_map_add_points(.base_map, assessment_data,
  gps.lng.column = "X_gps_longitude",
  gps.lat.column = "X_gps_latitude", color.by.column = "deviceid",
  color.group.column = "district_idp")
}
\arguments{
\item{.base_map}{a leaflet basemap, i.e. made with base_map()}

\item{assessment_data}{your assessment data as a data.frame (can use read.csv() to make this)}

\item{gps.lng.column}{name of the latitude column in quotes (optional)}

\item{color.by.column}{name of the column to color by (max 8 unique values)}

\item{color.group.column}{name of the column of groups between which colours don't need to be unique}

\item{gps.lon.column}{name of the latitude column in quotes (optional)}
}
\description{
add points of an assesment to the web map
}
