% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_cities.R
\docType{data}
\name{us_cities}
\alias{us_cities}
\title{US Major Cities}
\format{
A data frame with columns:
\describe{
\item{\code{name}}{City name}
\item{\code{capital}}{Capital status code indicates whether a city is a capital or not.
A value of 0 indicates that the city is not a capital,
while a value of 1 indicates that the city is a capital.
If the city is a state capital, the value is 2.}
\item{\code{lng}}{Longitude in decimal degrees.}
\item{\code{lat}}{Latitude in decimal degrees.}
}
}
\source{
The census-designated place population as of January 30, 2024,
as enumerated by the 2020 United States census.
}
\usage{
us_cities
}
\description{
This dataset contains the locations of cities within the United States
with populations of about 40,000 or greater, all state capitals, and the national capital.
}
\examples{
str(us_cities)
}
\keyword{datasets}
